OkPkgTemplates

Documentation for OkPkgTemplates.

OkPkgTemplates.chkdest

OkPkgTemplates.copymyfiles_script

OkPkgTemplates.mypkgtemplate_dir

OkPkgTemplates.pkgtemplating_script

OkPkgTemplates.stringscore

OkPkgTemplates.updateprojtoml_script

OkPkgTemplates.@genpkg

OkPkgTemplates.@upactions

OkPkgTemplates.chkdest — Method

chkdest() return Pkg.devdir() if DEFAULT_DESTINATION() is empty.

OkPkgTemplates.copymyfiles_script — Method

Copy some files from repo0 to repo1. List of files:

All files under .github/workflows

OkPkgTemplates.mypkgtemplate_dir — Method

Return a path relative to the default template file directory (OkPkgTemplates/mypkgtemplates).

OkPkgTemplates.pkgtemplating_script — Method

pkgtemplating_script(dest, yourpkgname) returns the script of PkgTemplates (quote ... end) to be executed at the scope that the macro is called.

It

creates a package using PkgTemplates.Templates

add Documenter, CompatHelperLocal, Test into Project.toml

OkPkgTemplates.stringscore — Method

stringscore(str) gives an absolute score for a string; the order of score from large to small corresponds to the alphabetical order. For a string less than 100 characters, the score should not exceeds one.

Example

julia> stringscore.(["AAA", "Aa", "BCD"])
3-element Vector{Float64}:
 0.0656565
 0.06597
 0.06667680000000001

OkPkgTemplates.updateprojtoml_script — Method

After making the template successfully, add "Documenter", "CompatHelperLocal" to [extras] and [targets] as runtests.jl (may) use them.

updateprojtoml_script(dest, yourpkgname) creates script that projtoml_path = joinpath(dest, yourpkgname, "Project.toml") will be modified on execution.

It modify Project.toml by add [extras] and [targets] for the scope of Test.

OkPkgTemplates.@genpkg — Macro

genpkg(yourpkgname::String) generate your package using presets.

If OkPkgTemplates.DEFAULT_DESTINATION() returns an empty string, it use Pkg.devdir() at user's scope.

In julia REPL,

key in OkPkgTemplates.DEFAULT_ and [tab] to list all functions that returns default variables

key in OkPkgTemplates.PLUGIN_ and [tab] to list all functions that returns ::PkgTemplates.Plugin presets.

Redefine these functions in the local scope to assign the variable.

Example

To specify output destination, redefine DEFAULT_DESTINATION()

OkPkgTemplates.DEFAULT_DESTINATION() = pwd()

and then generate the Package

@genpkg "MyNewProject"

Note

Feel free to redefine OkPkgTemplates.DEFAULT_...

Warning

In the following cases, write a new macro of your own referencing @genpkg instead, or just use PkgTemplates as instructed:

If you have the thought to redefine OkPkgTemplates.PLUGIN_...

If you want to set PkgTemplates.user_view (for example)

Also see the helps

copymyfiles_script

updateprojtoml_script

OkPkgTemplates.@upactions — Macro

Replace Github Actions (all the files in .github/workflows) with the latest version that generated from OkPkgTemplates. Noted that julia enviroment should be activated in the current directory for your package in dev to update.

Warning

Make sure all your action files (all the files in .github/workflows) is under the control of git for safety.

Also see the helps

copymyfiles_script

updateprojtoml_script