The use of macro is confusing. Thus, I made an exemplary @sayhello3 to help me understand.

The use of genpkg

OkPkgTemplates.genpkg — Function

genpkg(yourpkgname::String, fs...) execute scripts fs one by one in user's scope using OkPkgTemplates's utilities.

Please define OkPkgTemplates.DEFAULT_** for changing defaults.

If OkPkgTemplates.DEFAULT_DESTINATION() returns an empty string, it use Pkg.devdir() at user's scope.

Tip

In julia REPL,

key in OkPkgTemplates.DEFAULT_ and [tab] to list all functions that returns default variables

key in OkPkgTemplates.PLUGIN_ and [tab] to list all functions that returns ::PkgTemplates.Plugin presets.

Redefine these functions in the local scope to assign the variable.

Example

To specify output destination, redefine DEFAULT_DESTINATION()

OkPkgTemplates.DEFAULT_DESTINATION() = pwd()

and then generate the Package

@genpkg "MyNewProject" OkReg

Note

Feel free to redefine OkPkgTemplates.DEFAULT_...

Warning

In the following cases, write a new macro of your own referencing @genpkg instead, or just use PkgTemplates as instructed:

If you have the thought to redefine OkPkgTemplates.PLUGIN_...

If you want to set PkgTemplates.user_view (for example)

Also see the helps

copymyfiles_script

updateprojtoml_script

Procedure to add a new method of @genpkg

OkPkgTemplates.TemplateIdentifier — Type

For developer,

1. add new struct XXX <: TemplateIdentifier end for every new @genpkg.

2. add a new templating expression, e.g.,

pkgtemplating_xxx(dest, yourpkgname) = quote
    t = Template(;
        user=DEFAULT_USERNAME(),
        dir=$dest,
        julia=DEFAULT_JULIAVER(),
        plugins=[
            Git(; manifest=false),
            PLUGIN_COMPATHELPER(),
            PLUGIN_GITHUBACTION(),
            Codecov(), # https://about.codecov.io/
            Documenter{GitHubActions}(),
            PkgTemplates.Readme(),
            PkgTemplates.TagBot(; changelog=tagbot_changelog()),
            PLUGIN_TEST(),
            RegisterAction()
        ]
    )

    t($yourpkgname)
end

3. add a new method for @genpkg macro

get_exprs(::Type{XXX}) = [pkgtemplating_xxx, updateprojtoml_script] # Expressions to be executed
macro genpkg(yourpkgname::String, xxx::Type{XXX})
    dest = chkdest()
    return genpkg(dest, yourpkgname, get_exprs(xxx)...)
end