Reference

OkPkgTemplates.commit_msg — Constant

Notice

Currently, you have to add

    client-payload: '{"commit_msg": "${{ github.event.commits[0].message }}"}'

in register.yml manually.

In register.yml:

with Options for name: Trigger next workflow:

client-payload: passes data from Workflow 1 to Workflow 2

Example:
with:
  client-payload: '{"ref": "${{ github.ref }}", "sha": "${{ github.sha }}"}'

Commit message can be obtained from both

${{ github.event.commits[0].message }} and

${{ github.event.head_commit.message }}

(Not sure) the latter one could be empty

(Not sure) error with the latter one could result from permission: content: false

It is possible to Iterating over github.event.commits or use join function to concatenate commits.

Alternative way to Get a commit using curl.

Also see github context for what others are available.

Noted that multiline string is not acceptable for json, you have to use e.g. "# hello\n second line" istead as the commit message where Project.toml version is raised.

Be aware if you have in several un-pushed commits that changes 'Project.toml', and push them at once, "commit_msg": "${{ github.event.commits[0].message will be the first commit and it is not necessarily the commit to be registered. UpdateOkReg register a version for this package at the latest commit being pushed.

Notice

As the latest commit you push is not necessary the commit you want to register, unexpected results will occurr if you push several commits among either of them 'Project.toml' is modified.

For example:

You have in several un-pushed commits that changes 'Project.toml', and push them at once, it register the version number of the lastest commit BUT client_payload (client-payload in register.yml) store and pass the information of the earliest commit in your push.

You add [compat] at the first commit with message "Fix compatibility", and increase the version number with message "New keyword args. mkpath for mv2dir ...". Both modify Project.toml, and the last one is registered but in release log it shows only "Fix compatibility".

If you want to automatically add commit message release note of the registered commit but you are too lazy to create pull requests since you are almost the only person working on YourPackage, please

push everything else before you increase the version number, and

push only the Project.toml with commit message you'd like to be adopted as release note.

OkPkgTemplates.logfile_msg — Constant

OkPkgTemplates.TemplateIdentifier — Type

Create new template:

Inside OkPkgTemplates/src,

add concrete struct of Type{<:TemplateIdentifier}.

add a new templating expression (optional)

add a new method get_exprs for XXX

# 1
struct XXX <: TemplateIdentifier end

# 2
pkgtemplating_xxx(yourpkgname) = quote
    t = Template(;
        user=DEFAULT_USERNAME(),
        dir=DEFAULT_DESTINATION(),
        julia=DEFAULT_JULIAVER(),
        plugins=[
            Git(; manifest=false),
            PLUGIN_COMPATHELPER(),
            PLUGIN_GITHUBACTION(),
            Codecov(), # https://about.codecov.io/
            Documenter{GitHubActions}(),
            PkgTemplates.Readme(),
            PkgTemplates.TagBot(; changelog=tagbot_changelog()),
            PLUGIN_TEST(),
            RegisterAction()
        ]
    )

    t($yourpkgname)
end

# 3
get_exprs(::Type{XXX}) = [pkgtemplating_xxx, updateprojtoml_script]

Use the new template

Generate: genpkg("HelloMyPkg", XXX)

Update: update(XXX)

OkPkgTemplates.PLUGIN_REGISTER — Method

PLUGIN_REGISTER() by default uses templates in /home/runner/work/OkPkgTemplates.jl/OkPkgTemplates.jl/mypkgtemplates/for_okregistry/github/workflows/register.yml.

In the default template, job attempts to register the latest pushed commit of YourPackage where Project.toml is modified. If success, TagBot.yml will be triggered. See also PLUGIN_TAGBOT.

If Project.toml is modified but version number unchanged, it simply failed and you don't have to worry about anything.

Known issue

According to the default "register.yml", in the workflow "Register Package to OkRegistry" both the repo of YourPackage and OkRegistry is checked out, and register YourPackage to OkRegistry using PAT ACCESS_OKREGISTRY token that must be provided in advance via your Github account. By default, actions/checkout checkout the commit of the pushed event.

After the current version is successfully registerd, TagBot will be triggered. However, if there is something wrong in tagging the previous version, that is a version exists in OkRegistry but failed to be tagged with TagBot, TagBot will stop at that version; and thus, succeeding versions cannot be tagged. See README for trouble shooting.

OkPkgTemplates.PLUGIN_TAGBOT — Method

PLUGIN_TAGBOT() returns an PkgTemplates.TagBot. You can redefine this function.

By default it add github.event.client_payload.logfile_msg as part of the release note. Please must read Known issue in the doc of PLUGIN_REGISTER().

Changelog

For variables that you can use in the changelog template, see line 174-184, tagbot/action/changelog.py and TagBot/Changelogs. Which are

"compare_url"

"custom"

"issues"

"package"

"previous_release"

"pulls"

"sha"

"version"

"version_url"

Also see

Default changelog template.

More information

see PkgTemplates/src/plugins/tagbot.jl.

If your registry is public, you don't need to provide token. For more information, see here

OkPkgTemplates.copymyfiles_script — Method

Copy some files from repo0 to repo1. List of files:

All files under .github/workflows

OkPkgTemplates.generate — Method

generate(args...) evaluate the expressions returned by genpkg, and it takes exactly the same argument as genpkg.

Example

Generate julia package at ./MyNewProject

generate("MyNewProject", GeneralReg)

Generate julia package at .julia/dev/MyNewProject

using Pkg
OkPkgTemplates.DEFAULT_DESTINATION() = Pkg.devdir()
generate("MyNewProject", OkReg)

Don't do this

generate(joinpath(Pkg.devdir(), "MolchanCB"), OkReg)

OkPkgTemplates.genpkg — Method

genpkg(yourpkgname, tp::Type{<:TemplateIdentifier}) call @chkdest, and fall back to genpkg(yourpkgname, get_exprs(tp)...).

OkPkgTemplates.genpkg — Method

genpkg(yourpkgname::String, fs...) execute scripts fs one by one in user's scope using OkPkgTemplates's utilities.

Please define OkPkgTemplates.DEFAULT_** for changing defaults.

If OkPkgTemplates.DEFAULT_DESTINATION() is not defined, it is defined as Pkg.devdir() at user's scope in the first use.

Tip

In julia REPL,

key in OkPkgTemplates.DEFAULT_ and [tab] to list all functions that returns default variables

key in OkPkgTemplates.PLUGIN_ and [tab] to list all functions that returns ::PkgTemplates.Plugin presets.

Redefine these functions in the local scope to assign the variable.

Example

To specify output destination, redefine DEFAULT_DESTINATION()

OkPkgTemplates.DEFAULT_DESTINATION() = pwd()

and then generate the Package

t = genpkg("MyNewProject", GeneralReg)
OkPkgTemplates.eval(t)

Resource

PkgTemplates/Saving-Templates:

PkgTemplates.user_view: Extending-Existing-Plugins

Also see the helps

copymyfiles_script

updateprojtoml_script

OkPkgTemplates.get_exprs — Method

OkPkgTemplates.mypkgtemplate_dir — Method

Return a path relative to the default template file directory (OkPkgTemplates/mypkgtemplates).

OkPkgTemplates.pkgtemplating_gereg — Method

pkgtemplating_gereg(yourpkgname) returns the script of PkgTemplates (quote ... end) for general registry.

OkPkgTemplates.pkgtemplating_okreg — Method

pkgtemplating_okreg(yourpkgname) returns the script of PkgTemplates (quote ... end) to be executed at the scope that the macro is called.

It

creates a package using PkgTemplates.Templates

add Documenter, CompatHelperLocal, Test into Project.toml

OkPkgTemplates.stringscore — Method

stringscore(str) gives an absolute score for a string; the order of score from large to small corresponds to the alphabetical order. For a string less than 100 characters, the score should not exceeds one.

Example

julia> stringscore.(["AAA", "Aa", "BCD"])
3-element Vector{Float64}:
 0.0656565
 0.06597
 0.06667680000000001

OkPkgTemplates.upactions — Method

upactions(mod::Module, TI::Type{<:TemplateIdentifier}) return expressions for updating CIs, referencing the path of mod.

Example

using MyPkgWhereCIToBeUpdated
exprs = upactions($MyPkgWhereCIToBeUpdated $GeneralReg)
for ex in exprs
    @eval(OkPkgTemplates, $ex)
end

Warning

Make sure all your action files (all the files in .github/workflows) is under the control of git for safety.

Also see the helps

copymyfiles_script

updateprojtoml_script

OkPkgTemplates.upactions — Method

upactions return expressions for updating CIs using upactions referencing the path of the current activated project environment.

OkPkgTemplates.update — Method

Overwrite/Update Github actions of the repository of the julia package in development; make sure the pacakge is under version control that the process can be reverted.

Example

Activate environment OkPkgTemplates:

pkg> activate .
(OkPkgTemplates) pkg> dev OkFiles
    Resolving package versions...
    Updating `~/.julia/dev/OkPkgTemplates/Project.toml`
    [b25a35d4] + OkFiles v0.4.1 `~/.julia/dev/OkFiles`
    Updating `~/.julia/dev/OkPkgTemplates/Manifest.toml`
    [b25a35d4] + OkFiles v0.4.1 `~/.julia/dev/OkFiles`

Update CIs in OkFiles:

julia> using OkPkgTemplates, OkFiles
julia> update(OkFiles, GeneralReg)

update execute expressions returned by upactions one-by-one.

Running update overwrite/update Github actions of the repository of the currently activated julia package; make sure they are under version control that the process can be reverted.

Example

Activate environment OkPkgTemplates:

pkg> activate .
(OkPkgTemplates) pkg>

Update CIs in OkPkgTemplates:

julia> using OkPkgTemplates
julia> update(GeneralReg)

OkPkgTemplates.updateprojtoml_script — Method

After making the template successfully, add "Documenter", "CompatHelperLocal" to [extras] and [targets] as runtests.jl (may) use them.

updateprojtoml_script(yourpkgname) creates expressions in which projtoml_path = joinpath(OkPkgTemplates.DEFAULT_DESTINATION(), yourpkgname, "Project.toml").

It modify Project.toml by add [extras] and [targets] for the scope of Test.

OkPkgTemplates.@chkdest — Macro

@chkdest check whether OkPkgTemplates.DEFAULT_DESTINATION() is defined. It define OkPkgTemplates.DEFAULT_DESTINATION() = Pkg.devdir() if DEFAULT_DESTINATION() is empty.