The use of macro is confusing. Thus, I made an exemplary @sayhello3 to help me understand.

macro returns expression, and the ex::Expr is executed in the scope just like @eval(mod::Module, ex) (mod can be Main).

ex::Expr uses the surrounded scope (lexical scope):

macro f45()
    x = 5
    return :(f(45))
end

@f45 (macro with 1 method)

f = sin
@f45

0.8509035245341184

f = tan
@f45

1.6197751905438615

Variable defined in expression won't contaminate the surrounded scope,

julia> @f451.6197751905438615
julia> xERROR: UndefVarError: `x` not defined

unless

module Hello
    macro f45brutal()
        Hello.x = 5
        return :(f(45))
    end
    Hello.x = 4.99
end

Main.Hello

julia> Hello.x4.99
julia> Hello.@f45brutalERROR: UndefVarError: `f` not defined
julia> Hello.x5

Noted that this is true for Main scope.

Tips

Multiline quote ... end is in fact Expr(:block, ex1, ex2, ...)

The say-hello function

OkPkgTemplates.sayhello3 — Function

sayhello3(name) = "Hello world! $name", an example to understand metaprogramming.

The say-hello macro built upon

OkPkgTemplates.@sayhello3 — Macro

macro sayhello3(str::String)
    return Expr(:block, :(sayhello3($str)))
end

Example

OkPkgTemplates.@sayhello3 "Bruce Willey"

# output
"Hello world! Bruce Willey"

macro sayhello3(ex::Expr)
    return Expr(:block, :(sayhello3($(last(ex.args)))))
end

Example: variable name is not defined in macro's scope

name = "Bruce Willey"

# output
"Bruce Willey"

julia> OkPkgTemplates.@sayhello3 $name
ERROR: UndefVarError: `name` not defined

Explain

name is passed as expression in this case, and it is not defined since macro dispatch is not based on the AST evaluated at runtime. Thus an error of UndefVarError is raised.

See Macros and dispatch for more information.

Example: a creative application

OkPkgTemplates.@sayhello3 name = "Bruce Willey"

# output
"Hello world! Bruce Willey"

Explain

In this example, the last argument of the expression name = "Bruce Willey" is "Bruce Willey".

Example 3: Use @eval if you want name be evaluated

@eval OkPkgTemplates.@sayhello3 $name

# output
"Hello world! Bruce Willey"

Explain

@eval evalute the expression @sayhello3 $name at runtime.

Thus, the argument for @sayhello3 is evaluated as String, and @sayhello3(str::String) is the dispatched method.

macro sayhello3(symb::Symbol)
    return Expr(:block, :(sayhello3($(string(symb)))))
end

Example: A single the runtime variable is rendered as Symbol

name = "Bruce Willey"
OkPkgTemplates.@sayhello3 name

# output
"Hello world! name"

The whereami example

Macro call

OkPkgTemplates.@whereami — Macro

macro whereami()
    path_parse_time = "Hello/World" # parse time
    println("Destination defined at the parse time: $(path_parse_time)")
    return quote
        path_runtime = chkdest()
        println("Destination defined at the runtime: $(path_runtime)")
        OkPkgTemplates.DEFAULT_DESTINATION() = $path_parse_time
    end
end

julia> using OkPkgTemplates
julia> OkPkgTemplates.DEFAULT_DESTINATION()""
julia> OkPkgTemplates.@whereamiDestination defined at the parse time: Hello/World
Destination defined at the runtime: /home/runner/.julia/dev
julia> OkPkgTemplates.DEFAULT_DESTINATION()"Hello/World"

Function call

OkPkgTemplates.whereami — Function

function whereami()
    # OkPkgTemplates.DEFAULT_DESTINATION() = dest
    # # This is not allowed at the compile time
    path_parse_time = chkdest()

    ex = quote
        OkPkgTemplates.DEFAULT_DESTINATION() = $path_parse_time
    end
    @eval(OkPkgTemplates, $ex)
    return path_parse_time
end

using OkPkgTemplates
OkPkgTemplates.DEFAULT_DESTINATION()

""

OkPkgTemplates.whereami()

"/home/runner/.julia/dev"

OkPkgTemplates.DEFAULT_DESTINATION()

"/home/runner/.julia/dev"