Interface

generate

OkPkgTemplates.generate — Function

generate(args...) evaluate the expressions returned by genpkg, and it takes exactly the same argument as genpkg.

Example

Generate julia package at ./MyNewProject

generate("MyNewProject", GeneralReg)

Generate julia package at .julia/dev/MyNewProject

using Pkg
OkPkgTemplates.DEFAULT_DESTINATION() = Pkg.devdir()
generate("MyNewProject", OkReg)

Don't do this

generate(joinpath(Pkg.devdir(), "MolchanCB"), OkReg)

update

OkPkgTemplates.update — Function

Overwrite/Update Github actions of the repository of the julia package in development; make sure the pacakge is under version control that the process can be reverted.

Example

Activate environment OkPkgTemplates:

pkg> activate .
(OkPkgTemplates) pkg> dev OkFiles
    Resolving package versions...
    Updating `~/.julia/dev/OkPkgTemplates/Project.toml`
    [b25a35d4] + OkFiles v0.4.1 `~/.julia/dev/OkFiles`
    Updating `~/.julia/dev/OkPkgTemplates/Manifest.toml`
    [b25a35d4] + OkFiles v0.4.1 `~/.julia/dev/OkFiles`

Update CIs in OkFiles:

julia> using OkPkgTemplates, OkFiles
julia> update(OkFiles, GeneralReg)

update execute expressions returned by upactions one-by-one.

Running update overwrite/update Github actions of the repository of the currently activated julia package; make sure they are under version control that the process can be reverted.

Example

Activate environment OkPkgTemplates:

pkg> activate .
(OkPkgTemplates) pkg>

Update CIs in OkPkgTemplates:

julia> using OkPkgTemplates
julia> update(GeneralReg)

List of available TemplateIdentifier

GeneralReg

OkReg