OkPkgTemplates

Documentation for OkPkgTemplates.

OkPkgTemplates.chkdest

OkPkgTemplates.mypkgtemplate_dir

OkPkgTemplates.@genpkg

OkPkgTemplates.chkdest — Method

chkdest() return Pkg.devdir() if DEFAULT_DESTINATION() is empty.

OkPkgTemplates.mypkgtemplate_dir — Method

Return a path relative to the default template file directory (Shorthands/mypkgtemplates).

OkPkgTemplates.@genpkg — Macro

genpkg(yourpkgname::String) generate your package using presets.

If OkPkgTemplates.DEFAULT_DESTINATION() returns an empty string, it use Pkg.devdir() at user's scope.

In julia REPL,

key in OkPkgTemplates.DEFAULT_ and [tab] to list all functions that returns default variables

key in OkPkgTemplates.PLUGIN_ and [tab] to list all functions that returns ::PkgTemplates.Plugin presets.

Redefine these functions in the local scope to assign the variable.

Example

To specify output destination, redefine DEFAULT_DESTINATION()

OkPkgTemplates.DEFAULT_DESTINATION() = pwd()

and then generate the Package

@genpkg "MyNewProject"

Note

Feel free to redefine OkPkgTemplates.DEFAULT_...

Warning

In the following cases, write a new macro of your own referencing @genpkg instead, or just use PkgTemplates as instructed:

If you have the thought to redefine OkPkgTemplates.PLUGIN_...

If you want to set PkgTemplates.user_view (for example)