OkPkgTemplates

Documentation for OkPkgTemplates.

OkPkgTemplates.commit_msg

OkPkgTemplates.logfile_msg

OkPkgTemplates.TemplateIdentifier

OkPkgTemplates.PLUGIN_REGISTER

OkPkgTemplates.PLUGIN_TAGBOT

OkPkgTemplates.chkdest

OkPkgTemplates.copymyfiles_script

OkPkgTemplates.genpkg

OkPkgTemplates.genpkg

OkPkgTemplates.mypkgtemplate_dir

OkPkgTemplates.pkgtemplating_gereg

OkPkgTemplates.pkgtemplating_okreg

OkPkgTemplates.sayhello3

OkPkgTemplates.sayhello3

OkPkgTemplates.stringscore

OkPkgTemplates.upactions

OkPkgTemplates.updateprojtoml_script

OkPkgTemplates.@sayhello3

OkPkgTemplates.@sayhello3

OkPkgTemplates.@sayhello3

OkPkgTemplates.@sayhello3

OkPkgTemplates.@upactions

OkPkgTemplates.@upactions

OkPkgTemplates.commit_msg — Constant

Notice

Currently, you have to add

    client-payload: '{"commit_msg": "${{ github.event.commits[0].message }}"}'

in register.yml manually.

In register.yml:

with Options for name: Trigger next workflow:

client-payload: passes data from Workflow 1 to Workflow 2

Example:
with:
  client-payload: '{"ref": "${{ github.ref }}", "sha": "${{ github.sha }}"}'

Commit message can be obtained from both

${{ github.event.commits[0].message }} and

${{ github.event.head_commit.message }}

(Not sure) the latter one could be empty

(Not sure) error with the latter one could result from permission: content: false

It is possible to Iterating over github.event.commits or use join function to concatenate commits.

Alternative way to Get a commit using curl.

Also see github context for what others are available.

Noted that multiline string is not acceptable for json, you have to use e.g. "# hello\n second line" istead as the commit message where Project.toml version is raised.

Be aware if you have in several un-pushed commits that changes 'Project.toml', and push them at once, "commit_msg": "${{ github.event.commits[0].message will be the first commit and it is not necessarily the commit to be registered. UpdateOkReg register a version for this package at the latest commit being pushed.

Notice

As the latest commit you push is not necessary the commit you want to register, unexpected results will occurr if you push several commits among either of them 'Project.toml' is modified.

For example:

You have in several un-pushed commits that changes 'Project.toml', and push them at once, it register the version number of the lastest commit BUT client_payload (client-payload in register.yml) store and pass the information of the earliest commit in your push.

You add [compat] at the first commit with message "Fix compatibility", and increase the version number with message "New keyword args. mkpath for mv2dir ...". Both modify Project.toml, and the last one is registered but in release log it shows only "Fix compatibility".

If you want to automatically add commit message release note of the registered commit but you are too lazy to create pull requests since you are almost the only person working on YourPackage, please

push everything else before you increase the version number, and

push only the Project.toml with commit message you'd like to be adopted as release note.

OkPkgTemplates.logfile_msg — Constant

OkPkgTemplates.PLUGIN_REGISTER — Method

PLUGIN_REGISTER() by default uses templates in /home/runner/work/OkPkgTemplates.jl/OkPkgTemplates.jl/mypkgtemplates/for_okregistry/github/workflows/register.yml.

In the default template, job attempts to register the latest pushed commit of YourPackage where Project.toml is modified. If success, TagBot.yml will be triggered. See also PLUGIN_TAGBOT.

If Project.toml is modified but version number unchanged, it simply failed and you don't have to worry about anything.

Known issue

According to the default "register.yml", in the workflow "Register Package to OkRegistry" both the repo of YourPackage and OkRegistry is checked out, and register YourPackage to OkRegistry using PAT ACCESS_OKREGISTRY token that must be provided in advance via your Github account. By default, actions/checkout checkout the commit of the pushed event.

After the current version is successfully registerd, TagBot will be triggered. However, if there is something wrong in tagging the previous version, that is a version exists in OkRegistry but failed to be tagged with TagBot, TagBot will stop at that version; and thus, succeeding versions cannot be tagged. See README for trouble shooting.

OkPkgTemplates.PLUGIN_TAGBOT — Method

PLUGIN_TAGBOT() returns an PkgTemplates.TagBot. You can redefine this function.

By default it add github.event.client_payload.logfile_msg as part of the release note. Please must read Known issue in the doc of PLUGIN_REGISTER().

Changelog

For variables that you can use in the changelog template, see line 174-184, tagbot/action/changelog.py and TagBot/Changelogs. Which are

"compare_url"

"custom"

"issues"

"package"

"previous_release"

"pulls"

"sha"

"version"

"version_url"

Also see

Default changelog template.

More information

see PkgTemplates/src/plugins/tagbot.jl.

If your registry is public, you don't need to provide token. For more information, see here

OkPkgTemplates.chkdest — Method

chkdest() return Pkg.devdir() if DEFAULT_DESTINATION() is empty.

OkPkgTemplates.copymyfiles_script — Method

Copy some files from repo0 to repo1. List of files:

All files under .github/workflows

OkPkgTemplates.genpkg — Method

genpkg(yourpkgname::String, fs...) execute scripts fs one by one in user's scope using OkPkgTemplates's utilities.

Please define OkPkgTemplates.DEFAULT_** for changing defaults.

If OkPkgTemplates.DEFAULT_DESTINATION() returns an empty string, it use Pkg.devdir() at user's scope.

Tip

In julia REPL,

key in OkPkgTemplates.DEFAULT_ and [tab] to list all functions that returns default variables

key in OkPkgTemplates.PLUGIN_ and [tab] to list all functions that returns ::PkgTemplates.Plugin presets.

Redefine these functions in the local scope to assign the variable.

Example

To specify output destination, redefine DEFAULT_DESTINATION()

OkPkgTemplates.DEFAULT_DESTINATION() = pwd()

and then generate the Package

@genpkg "MyNewProject" OkReg

Note

Feel free to redefine OkPkgTemplates.DEFAULT_...

Warning

In the following cases, write a new macro of your own referencing @genpkg instead, or just use PkgTemplates as instructed:

If you have the thought to redefine OkPkgTemplates.PLUGIN_...

If you want to set PkgTemplates.user_view (for example)

Also see the helps

copymyfiles_script

updateprojtoml_script

OkPkgTemplates.mypkgtemplate_dir — Method

Return a path relative to the default template file directory (OkPkgTemplates/mypkgtemplates).

OkPkgTemplates.pkgtemplating_gereg — Method

pkgtemplating_gereg(dest, yourpkgname) returns the script of PkgTemplates (quote ... end) for general registry.

OkPkgTemplates.pkgtemplating_okreg — Method

pkgtemplating_okreg(dest, yourpkgname) returns the script of PkgTemplates (quote ... end) to be executed at the scope that the macro is called.

It

creates a package using PkgTemplates.Templates

add Documenter, CompatHelperLocal, Test into Project.toml

OkPkgTemplates.sayhello3 — Method

sayhello3(name) = "Hello world! $name", an example to understand metaprogramming.

OkPkgTemplates.stringscore — Method

stringscore(str) gives an absolute score for a string; the order of score from large to small corresponds to the alphabetical order. For a string less than 100 characters, the score should not exceeds one.

Example

julia> stringscore.(["AAA", "Aa", "BCD"])
3-element Vector{Float64}:
 0.0656565
 0.06597
 0.06667680000000001

OkPkgTemplates.upactions — Method

upactions(pwd1::String, pkgname::String, TI::Type{<:TemplateIdentifier}) return expression for replacing Github Actions (all the files in .github/workflows) with the latest version that generated from OkPkgTemplates.

This function is separated for test purpose. For user, please use @upactions.

Example

ex = OkPkgTemplates.upactions(dir_targetfolder(), pkgname2build)
@eval(OkPkgTemplates, $ex) # expression must be evaluated under the scope of OkPkgTemplates; otherwise, error will occur since the current scope might not have pakage required in `ex`.

Warning

Make sure all your action files (all the files in .github/workflows) is under the control of git for safety.

Also see the helps

copymyfiles_script

updateprojtoml_script

OkPkgTemplates.updateprojtoml_script — Method

After making the template successfully, add "Documenter", "CompatHelperLocal" to [extras] and [targets] as runtests.jl (may) use them.

updateprojtoml_script(dest, yourpkgname) creates script that projtoml_path = joinpath(dest, yourpkgname, "Project.toml") will be modified on execution.

It modify Project.toml by add [extras] and [targets] for the scope of Test.

OkPkgTemplates.@sayhello3 — Macro

macro sayhello3(ex::Expr)
    return Expr(:block, :(sayhello3($(last(ex.args)))))
end

Example: variable name is not defined in macro's scope

name = "Bruce Willey"

# output
"Bruce Willey"

julia> OkPkgTemplates.@sayhello3 $name
ERROR: UndefVarError: `name` not defined

Explain

name is passed as expression in this case, and it is not defined since macro dispatch is not based on the AST evaluated at runtime. Thus an error of UndefVarError is raised.

See Macros and dispatch for more information.

Example: a creative application

OkPkgTemplates.@sayhello3 name = "Bruce Willey"

# output
"Hello world! Bruce Willey"

Explain

In this example, the last argument of the expression name = "Bruce Willey" is "Bruce Willey".

Example 3: Use @eval if you want name be evaluated

@eval OkPkgTemplates.@sayhello3 $name

# output
"Hello world! Bruce Willey"

Explain

@eval evalute the expression @sayhello3 $name at runtime.

Thus, the argument for @sayhello3 is evaluated as String, and @sayhello3(str::String) is the dispatched method.

OkPkgTemplates.@sayhello3 — Macro

macro sayhello3(str::String)
    return Expr(:block, :(sayhello3($str)))
end

Example

OkPkgTemplates.@sayhello3 "Bruce Willey"

# output
"Hello world! Bruce Willey"

OkPkgTemplates.@sayhello3 — Macro

macro sayhello3(symb::Symbol)
    return Expr(:block, :(sayhello3($(string(symb)))))
end

Example: A single the runtime variable is rendered as Symbol

name = "Bruce Willey"
OkPkgTemplates.@sayhello3 name

# output
"Hello world! name"

OkPkgTemplates.@upactions — Macro

@upactions(mod::Module, TI::Type{<:TemplateIdentifier}) update CIs using upactions referencing the path of mod. See upactions.

Example

using MyPkgWhereCIToBeUpdated
@eval @upactions $MyPkgWhereCIToBeUpdated $GeneralReg
# execute `@upactions` whth `GeneralReg` evaluated as `Type{<:TemplateIdentifier}` and `MyPkgWhereCIToBeUpdated` evaluated as `Module`.

OkPkgTemplates.@upactions — Macro

@upactions update CIs using upactions referencing the path of the current activated project environment. See upactions.

Example

@eval @upactions $GeneralReg
# execute `@upactions` whth `GeneralReg` evaluated as `Type{<:TemplateIdentifier}`